Initial Commit. Creates mentor training slides.
This commit adds the basis for the mentor training slides. Some
additional formatting will still need to be done, but all of the
necessary content is there.

This set of slides will serve as the basis for other presentations,
which will hopefully contain solely markdown files (rather than raw
HTML, CSS, and JavaScript) at some point, and will be converted by a
build system to the final rendered pages.